rename package from DocumentWorker to task4DocumentWorker
